requirements.txt
streamlit
streamlit-extras
pandas
snowflake-connector-python
snowflake-snowpark-python